+++ src/beetle-ui/src/Route/Home.elm
@@ -4,11 +4,11 @@ import Button
 import Dict
 import Environment
 import Html
-import Html.Attributes
+import Html.Attributes as A
 import Html.Events
 import Http
 import Icon
-import Json.Decode
+import Json.Decode as D
 import Json.Encode
 import Random
 
@@ -18,14 +18,19 @@ type Alert
     | Happy String
 
 
+type alias DeviceSnapshot =
+    { nickname : Maybe String }
+
+
 type alias OwnedDevice =
     { id : String
+    , nickname : Maybe String
     , busy : Bool
     }
 
 
 type alias Data =
-    { devices : List OwnedDevice
+    { devices : Result Http.Error (List OwnedDevice)
     , newDevice : ( String, Maybe (Maybe Http.Error) )
     , alert : Maybe Alert
     }
@@ -36,8 +41,7 @@ type alias Model =
 
 
 type alias DeviceList =
-    { devices : Dict.Dict String Int
-    }
+    { devices : Dict.Dict String DeviceSnapshot }
 
 
 type Message
@@ -55,7 +59,7 @@ type alias RegistrationResponse =
 
 emptyData : Data
 emptyData =
-    { devices = [], newDevice = ( "", Nothing ), alert = Nothing }
+    { devices = Ok [], newDevice = ( "", Nothing ), alert = Nothing }
 
 
 hasPendingAddition : Data -> Bool
@@ -72,9 +76,9 @@ hasPendingAddition data =
             False
 
 
-registrationDecoder : Json.Decode.Decoder RegistrationResponse
+registrationDecoder : D.Decoder RegistrationResponse
 registrationDecoder =
-    Json.Decode.map RegistrationResponse (Json.Decode.field "id" Json.Decode.string)
+    D.map RegistrationResponse (D.field "id" D.string)
 
 
 removeDevice : Environment.Environment -> String -> Cmd Message
@@ -161,7 +165,7 @@ setNewDeviceId id model =
 
 ownedDevice : String -> OwnedDevice
 ownedDevice id =
-    OwnedDevice id False
+    { id = id, nickname = Nothing, busy = False }
 
 
 checkBusy : String -> OwnedDevice -> OwnedDevice
@@ -171,7 +175,21 @@ checkBusy id dev =
 
 markDeviceBusy : String -> Data -> Data
 markDeviceBusy id data =
-    { data | devices = List.map (checkBusy id) data.devices }
+    { data | devices = Result.map (List.map (checkBusy id)) data.devices }
+
+
+ownedDeviceFromKeyValue : ( String, DeviceSnapshot ) -> OwnedDevice
+ownedDeviceFromKeyValue pair =
+    let
+        ( id, snapshot ) =
+            pair
+    in
+    { id = id, busy = False, nickname = snapshot.nickname }
+
+
+deviceListFromMap : Dict.Dict String DeviceSnapshot -> List OwnedDevice
+deviceListFromMap mapping =
+    Dict.toList mapping |> List.map ownedDeviceFromKeyValue
 
 
 update : Environment.Environment -> Message -> Model -> ( Model, Cmd Message )
@@ -179,22 +197,31 @@ update env message model =
     case message of
         LoadedDevices deviceResponse ->
             case ( deviceResponse, model ) of
-                ( Err e, _ ) ->
-                    ( Just (Ok emptyData), Cmd.none )
+                ( Err e, maybeModel ) ->
+                    let
+                        newModel =
+                            case maybeModel of
+                                Just (Ok m) ->
+                                    { m | devices = Err e }
+
+                                _ ->
+                                    { emptyData | devices = Err e }
+                    in
+                    ( Just (Ok newModel), Cmd.none )
 
                 ( Ok responseData, Just (Ok loadedModel) ) ->
                     let
                         devices =
-                            Dict.keys responseData.devices |> List.map ownedDevice
+                            deviceListFromMap responseData.devices
                     in
-                    ( Just (Ok { loadedModel | devices = devices }), Cmd.none )
+                    ( Just (Ok { loadedModel | devices = Ok devices }), Cmd.none )
 
                 ( Ok responseData, _ ) ->
                     let
                         devices =
-                            Dict.keys responseData.devices |> List.map ownedDevice
+                            deviceListFromMap responseData.devices
                     in
-                    ( Just (Ok { emptyData | devices = devices }), Cmd.none )
+                    ( Just (Ok { emptyData | devices = Ok devices }), Cmd.none )
 
         AttemptDeviceClaim ->
             let
@@ -223,14 +250,26 @@ update env message model =
 
 renderDevice : Environment.Environment -> OwnedDevice -> Html.Html Message
 renderDevice env device =
+    let
+        linkContents =
+            case device.nickname of
+                Just name ->
+                    Html.div [ A.title device.id ] [ Html.text name ]
+
+                Nothing ->
+                    Html.div [] [ Html.text device.id ]
+
+        linkUrl =
+            Environment.buildRoutePath env ("devices/" ++ device.id)
+    in
     Html.tr []
         [ Html.td
-            [ Html.Attributes.class "px-3 py-2" ]
+            [ A.class "px-3 py-2" ]
             [ Html.a
-                [ Html.Attributes.href (Environment.buildRoutePath env ("devices/" ++ device.id)) ]
-                [ Html.text device.id ]
+                [ A.href linkUrl ]
+                [ linkContents ]
             ]
-        , Html.td [ Html.Attributes.class "px-3 py-2 text-right" ]
+        , Html.td [ A.class "px-3 py-2 text-right" ]
             (if device.busy then
                 [ Button.view (Button.DisabledIcon Icon.Trash) ]
 
@@ -245,17 +284,25 @@ deviceList data env =
     let
         addButton =
             Button.view (Button.LinkIcon Icon.Add (Environment.buildRoutePath env "register-device"))
+
+        body =
+            case data.devices of
+                Ok list ->
+                    Html.tbody [] (List.map (renderDevice env) list)
+
+                Err error ->
+                    Html.tbody [] [ Html.tr [] [ Html.td [] [ Html.text "Failed to load" ] ] ]
     in
-    Html.div [ Html.Attributes.class "flex-1" ]
-        [ Html.table [ Html.Attributes.class "w-full" ]
+    Html.div [ A.class "flex-1" ]
+        [ Html.table [ A.class "w-full" ]
             [ Html.thead []
-                [ Html.tr [ Html.Attributes.class "text-left" ]
-                    [ Html.th [ Html.Attributes.class "px-3 pb-2" ] [ Html.text "Devices" ]
-                    , Html.th [ Html.Attributes.class "px-3 py-2 text-right" ]
+                [ Html.tr [ A.class "text-left" ]
+                    [ Html.th [ A.class "px-3 pb-2" ] [ Html.text "Devices" ]
+                    , Html.th [ A.class "px-3 py-2 text-right" ]
                         [ addButton ]
                     ]
                 ]
-            , Html.tbody [] (List.map (renderDevice env) data.devices)
+            , body
             ]
         ]
 
@@ -264,22 +311,32 @@ view : Model -> Environment.Environment -> Html.Html Message
 view model env =
     case model of
         Nothing ->
-            Html.div [ Html.Attributes.class "flex px-4 py-3" ] [ Html.text "Loading..." ]
+            Html.div [ A.class "flex px-4 py-3" ] [ Html.text "Loading..." ]
 
         Just result ->
             case result of
                 Err error ->
-                    Html.div [ Html.Attributes.class "flex px-4 py-3" ]
+                    Html.div [ A.class "flex px-4 py-3" ]
                         [ Html.text "failed, please refresh" ]
 
                 Ok modelData ->
-                    Html.div [ Html.Attributes.class "flex px-4 py-3" ]
+                    Html.div [ A.class "flex px-4 py-3" ]
                         [ deviceList modelData env ]
 
 
-sessionDecoder : Json.Decode.Decoder DeviceList
+sessionDecoder : D.Decoder DeviceList
 sessionDecoder =
-    Json.Decode.map DeviceList (Json.Decode.field "devices" (Json.Decode.dict Json.Decode.int))
+    D.map DeviceList (D.field "devices" (D.dict snapshotDecoder))
+
+
+
+-- TODO(API): the api returns devices in the same payload as other user information. That will probably
+-- not paginate well.
+
+
+snapshotDecoder : D.Decoder DeviceSnapshot
+snapshotDecoder =
+    D.map DeviceSnapshot (D.field "nickname" (D.maybe D.string))
 
 
 fetchDevices : Environment.Environment -> Cmd Message
